+++ src/Compiler/IL/TempLocalAllocator.cs
@@ -688,6 +688,10 @@ internal static class TempLocalAllocator
                 yield return getItemAsNumber.Object;
                 yield return getItemAsNumber.Index;
                 break;
+            case LIRGetItemAsNumberString getItemAsNumberString:
+                yield return getItemAsNumberString.Object;
+                yield return getItemAsNumberString.Index;
+                break;
             case LIRSetItem setItem:
                 yield return setItem.Object;
                 yield return setItem.Index;
@@ -1148,6 +1152,9 @@ internal static class TempLocalAllocator
                 return true;
             case LIRGetItemAsNumber getItemAsNumber:
                 defined = getItemAsNumber.Result;
+                return true;
+            case LIRGetItemAsNumberString getItemAsNumberString:
+                defined = getItemAsNumberString.Result;
                 return true;
 
             case LIRSetItem setItem: